XStringPartialMatches-class {Biostrings}	R Documentation

XStringPartialMatches objects

Description

WARNING: This class is currently under development and might not work properly! Full documentation will come later.

Please DO NOT TRY TO USE it for now. Thanks for your comprehension!

Accessor methods

In the code snippets below, x is an XStringPartialMatches object.

subpatterns(x): Not ready yet.

pattern(x): Not ready yet.

Standard generic methods

In the code snippets below, x is an XStringPartialMatches objects, and i can be a numeric or logical vector.

x[i]: Return a new XStringPartialMatches object made of the selected views. i can be a numeric vector, a logical vector, NULL or missing. The returned object has the same subject as x.

Author(s)

H. Pages

See Also

XStringViews-class, XString-class, letter

[Package Biostrings version 2.32.0 Index]