detail {Biostrings}	R Documentation

Show (display) detailed object content

Description

This is a variant of show, offering a more detailed display of object content.

Usage

detail(x, ...)

Arguments

x
An object. The default simply invokes show.

...
Additional arguments. The default definition makes no use of these arguments.

Value

None; the function is invoked for its side effect (detailed display of object content).

Author(s)

Martin Morgan

Examples

origMAlign <-
  readDNAMultipleAlignment(filepath =
                           system.file("extdata",
                                       "msx2_mRNA.aln",
                                       package="Biostrings"),
                           format="clustal")
detail(origMAlign)

[Package Biostrings version 2.32.0 Index]