getSeq {Biostrings}	R Documentation

getSeq

Description

A generic function for extracting a set of sequences (or subsequences) from a sequence container like a BSgenome object or other.

Usage

getSeq(x, ...)

Arguments

x
A BSgenome object or any other supported object. Do showMethods("getSeq") to get the list of all supported types for x.

...
Any additional arguments needed by the specialized methods.

Value

An XString object or an XStringSet object or a character vector containing the extracted sequence(s).

See man pages of individual methods for the details e.g. with ?`getSeq,BSgenome-method` to access the man page of the method for BSgenome objects (make sure the BSgenome package is loaded first).

See Also

getSeq,BSgenome-method, XString-class, XStringSet-class

Examples

## Note that you need to load the package(s) defining the specialized
## methods to have showMethods() display them and to be able to access
## their man pages:
library(BSgenome)
showMethods("getSeq")

[Package Biostrings version 2.32.0 Index]