longestConsecutive {Biostrings}	R Documentation

Obtain the length of the longest substring containing only 'letter'

Description

This function accepts a character vector and computes the length of the longest substring containing only letter for each element of x.

Usage

longestConsecutive(seq, letter) 

Arguments

seq
Character vector.

letter
Character vector of length 1, containing one single character.

Details

The elements of x can be in upper case, lower case or mixed. NAs are handled.

Value

An integer vector of the same length as x.

Author(s)

W. Huber

Examples

 v = c("AAACTGTGFG", "GGGAATT", "CCAAAAAAAAAATT")
 longestConsecutive(v, "A")

[Package Biostrings version 2.50.1 Index]