matchProbePair {Biostrings}	R Documentation

Find "theoretical amplicons" mapped to a probe pair

Description

In the context of a computer-simulated PCR experiment, one wants to find the amplicons mapped to a given primer pair. The matchProbePair function can be used for this: given a forward and a reverse probe (i.e. the chromosome-specific sequences of the forward and reverse primers used for the experiment) and a target sequence (generally a chromosome sequence), the matchProbePair function will return all the "theoretical amplicons" mapped to this probe pair.

Usage

matchProbePair(Fprobe, Rprobe, subject, algorithm="auto", logfile=NULL, verbose=FALSE)

Arguments

Fprobe
The forward probe.

Rprobe
The reverse probe.

subject
A DNAString object (or an XStringViews object with a DNAString subject) containing the target sequence.

algorithm
One of the following: "auto", "naive-exact", "naive-inexact", "boyer-moore" or "shift-or". See matchPattern for more information.

logfile
A file used for logging.

verbose
TRUE or FALSE.

Details

The matchProbePair function does the following: (1) find all the "plus hits" i.e. the Fprobe and Rprobe matches on the "plus" strand, (2) find all the "minus hits" i.e. the Fprobe and Rprobe matches on the "minus" strand and (3) from the set of all (plus\_hit, minus\_hit) pairs, extract and return the subset of "reduced matches" i.e. the (plus\_hit, minus\_hit) pairs such that (a) plus\_hit <= minus\_hit and (b) there are no hits (plus or minus) between plus\_hit and minus\_hit. This set of "reduced matches" is the set of "theoretical amplicons".

Value

An XStringViews object containing the set of "theoretical amplicons".

Author(s)

H. Pages

See Also

matchPattern, matchLRPatterns, findPalindromes, reverseComplement, XStringViews-class

Examples

  library(BSgenome.Dmelanogaster.UCSC.dm3)
  subject <- Dmelanogaster$chr3R

  ## With 20-nucleotide forward and reverse probes:
  Fprobe <- "AGCTCCGAGTTCCTGCAATA"
  Rprobe <- "CGTTGTTCACAAATATGCGG"
  matchProbePair(Fprobe, Rprobe, subject) # 1 "theoretical amplicon"

  ## With shorter forward and reverse probes, the risk of having multiple
  ## "theoretical amplicons" increases:
  Fprobe <- "AGCTCCGAGTTCC"
  Rprobe <- "CGTTGTTCACAA"
  matchProbePair(Fprobe, Rprobe, subject) # 2 "theoretical amplicons"
  Fprobe <- "AGCTCCGAGTT"
  Rprobe <- "CGTTGTTCACA"
  matchProbePair(Fprobe, Rprobe, subject) # 9 "theoretical amplicons"

[Package Biostrings version 2.32.0 Index]