phiX174Phage {Biostrings}	R Documentation

Versions of bacteriophage phiX174 complete genome and sample short reads

Description

Six versions of the complete genome for bacteriophage φ X174 as well as a small number of Solexa short reads, qualities associated with those short reads, and counts for the number times those short reads occurred.

Details

The phiX174Phage object is a DNAStringSet containing the following six naturally occurring versions of the bacteriophage φ X174 genome cited in Smith et al.:

Genbank:

The version of the genome from GenBank (NC\_001422.1, GI:9626372).

RF70s:

A preparation of φ X double-stranded replicative form (RF) of DNA by Clyde A. Hutchison III from the late 1970s.

SS78:

A preparation of φ X virion single-stranded DNA from 1978.

Bull:

The sequence of wild-type φ X used by Bull et al.

G'97:

The φ X replicative form (RF) of DNA from Bull et al.

NEB'03:

A φ X replicative form (RF) of DNA from New England BioLabs (NEB).

The srPhiX174 object is a DNAStringSet containing short reads from a Solexa machine.

The quPhiX174 object is a BStringSet containing Solexa quality scores associated with srPhiX174.

The wtPhiX174 object is an integer vector containing counts associated with srPhiX174.

References

http://www.genome.jp/dbget-bin/www_bget?refseq+NC_001422

Bull, J. J., Badgett, M. R., Wichman, H. A., Huelsenbeck, Hillis, D. M., Gulati, A., Ho, C. & Molineux, J. (1997) Genetics 147, 1497-1507.

Smith, Hamilton O.; Clyde A. Hutchison, Cynthia Pfannkoch, J. Craig Venter (2003-12-23). "Generating a synthetic genome by whole genome assembly: {phi}X174 bacteriophage from synthetic oligonucleotides". Proceedings of the National Academy of Sciences 100 (26): 15440-15445. doi:10.1073/pnas.2237126100.

Examples

data(phiX174Phage)
nchar(phiX174Phage)
genBankPhage <- phiX174Phage[[1]]
genBankSubstring <- substring(genBankPhage, 2793-34, 2811+34)

data(srPhiX174)
srPhiX174
quPhiX174
summary(wtPhiX174)

alignPhiX174 <-
  pairwiseAlignment(srPhiX174, genBankSubstring,
                    patternQuality = SolexaQuality(quPhiX174),
                    subjectQuality = SolexaQuality(99L),
                    type = "global-local")
summary(alignPhiX174, weight = wtPhiX174)

[Package Biostrings version 2.32.0 Index]