pmatchPattern {Biostrings}	R Documentation

Longest Common Prefix/Suffix/Substring searching functions

Description

Functions for searching the Longest Common Prefix/Suffix/Substring of two strings.

WARNING: These functions are experimental and might not work properly! Full documentation will come later.

Please send questions/comments to hpages@fredhutch.org

Thanks for your comprehension!

Usage

lcprefix(s1, s2)
lcsuffix(s1, s2)
lcsubstr(s1, s2)
pmatchPattern(pattern, subject, maxlength.out=1L)

Arguments

s1
1st string, a character string or an XString object.

s2
2nd string, a character string or an XString object.

pattern
The pattern string.

subject
An XString object containing the subject string.

maxlength.out
The maximum length of the output i.e. the maximum number of views in the returned object.

See Also

matchPattern, XStringViews-class, XString-class

[Package Biostrings version 2.50.1 Index]