toComplex {Biostrings}	R Documentation

Turning a DNA sequence into a vector of complex numbers

Description

The toComplex utility function turns a DNAString object into a complex vector.

Usage

toComplex(x, baseValues)

Arguments

x
A DNAString object.

baseValues
A named complex vector containing the values associated to each base e.g. c(A=1+0i, G=0+1i, T=-1+0i, C=0-1i)

Value

A complex vector of the same length as x.

Author(s)

H. Pagès

See Also

DNAString

Examples

  seq <- DNAString("accacctgaccattgtcct")
  baseValues1 <- c(A=1+0i, G=0+1i, T=-1+0i, C=0-1i)
  toComplex(seq, baseValues1)

  ## GC content:
  baseValues2 <- c(A=0, C=1, G=1, T=0)
  sum(as.integer(toComplex(seq, baseValues2)))
  ## Note that there are better ways to do this (see ?alphabetFrequency)

[Package Biostrings version 2.50.1 Index]